refactor: remove disabled & disabledValue of Picker
