+++ packages/apps/stock/src/components/templates/stockCorrection/StockCorrectionReasonPicker/StockCorrectionReasonPicker.js
@@ -57,8 +57,7 @@ const StockCorrectionReasonPicker = ({
       labelField="name"
       valueField="id"
       required={true}
-      disabled={status === StockCorrection.status.Validated}
-      disabledValue={reason?.name}
+      readonly={status === StockCorrection.status.Validated}
       isScrollViewContainer={true}
     />
   );
